Use a u8 instead of a u32, it's faster.

+++ src/binary_reader.rs
@@ -1239,7 +1239,7 @@ impl<'a> BinaryReader<'a> {
     }
 
     fn read_0xfd_operator(&mut self) -> Result<Operator<'a>> {
-        let code = self.read_var_u32()? as u32;
+        let code = self.read_var_u32()? as u8;
         Ok(match code {
             0x00 => Operator::V128Load {
                 memarg: self.read_memarg()?,
